Comentarios para nuevas modificaciones
