+++ src/main/java/cl/uchile/dcc/finalreality/model/character/player/MageCharacter.java
@@ -14,15 +14,15 @@ public interface MageCharacter extends PlayerCharacter {
   /**
    * Returns the maximum value of MP.
    */
-  public int getMaxMp();
+  int getMaxMp();
 
   /**
    * Sets the current MP of the character to {@code newMp}.
    */
-  public void setCurrentMp(int newMp) throws InvalidStatValueException;
+  void setCurrentMp(int newMp) throws InvalidStatValueException;
 
   /**
    * Returns the current MP of the character.
    */
-  public int getCurrentMp();
+  int getCurrentMp();
 }
