+++ src/main/java/cl/uchile/dcc/finalreality/model/character/player/PlayerCharacter.java
@@ -20,11 +20,6 @@ import cl.uchile.dcc.finalreality.model.weapon.Weapon;
  */
 public interface PlayerCharacter extends GameCharacter {
   /**
-   * Equips a weapon to a character when it is allowed.
-   */
-  // void equip(Weapon weapon);
-
-  /**
    * Return this character's equipped weapon.
    */
   Weapon getEquippedWeapon();
