+++ src/main/java/cl/uchile/dcc/finalreality/model/weapon/Bow.java
@@ -1,8 +1,8 @@
 package cl.uchile.dcc.finalreality.model.weapon;
 
 import cl.uchile.dcc.finalreality.model.character.player.PlayerCharacter;
-import cl.uchile.dcc.finalreality.model.weapon.ConditionInterfaces.EquippableByEngineer;
-import cl.uchile.dcc.finalreality.model.weapon.ConditionInterfaces.EquippableByThief;
+import cl.uchile.dcc.finalreality.model.weapon.interfaces.EquippableByEngineer;
+import cl.uchile.dcc.finalreality.model.weapon.interfaces.EquippableByThief;
 import java.util.Objects;
 import org.jetbrains.annotations.NotNull;
