--- src/main/java/cl/uchile/dcc/finalreality/model/weapon/ConditionInterfaces/EquippableByEngineer.java
@@ -1,7 +1,0 @@
-package cl.uchile.dcc.finalreality.model.weapon.ConditionInterfaces;
-
-import cl.uchile.dcc.finalreality.model.character.player.PlayerCharacter;
-
-public interface EquippableByEngineer {
-  void equipToEngineer(PlayerCharacter character);
-}
