--- src/main/java/cl/uchile/dcc/finalreality/model/weapon/ConditionInterfaces/EquippableByKnight.java
@@ -1,7 +1,0 @@
-package cl.uchile.dcc.finalreality.model.weapon.ConditionInterfaces;
-
-import cl.uchile.dcc.finalreality.model.character.player.PlayerCharacter;
-
-public interface EquippableByKnight {
-  void equipToKnight(PlayerCharacter character);
-}
