--- src/main/java/cl/uchile/dcc/finalreality/model/weapon/ConditionInterfaces/EquippableByThief.java
@@ -1,7 +1,0 @@
-package cl.uchile.dcc.finalreality.model.weapon.ConditionInterfaces;
-
-import cl.uchile.dcc.finalreality.model.character.player.PlayerCharacter;
-
-public interface EquippableByThief {
-  void equipToThief(PlayerCharacter character);
-}
